pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Patrick Brennan", email = "brennanpatrick20@gmail.com"},
]
dependencies = [
    "more-itertools>=9.0.0",
    "requests>=2.28.1",
    "python-dotenv>=0.21.0",
    "dateparser>=1.1.4",
    "tqdm>=4.64.1",
    "colorama>=0.4.6",
    "numpy>=1.23.5",
    "black>=22.12.0",
]
requires-python = ">=3.10"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.pdm.scripts]
new = "python -m aoc"
test = "python -m aoc --test"
submit = "python -m aoc --submit --part"
format = "black ."
[tool]
[tool.pdm]
